requirements.txt
opencv-python==4.5.4.60
grpcio
grpcio-tools